AiThGameLocalizer/requirements.txt
flask
flask-cors
pywebview
llama-cpp-python
huggingface_hub
pyinstaller